The terminal state is now reset before displaying the disconnect or taskwindow closure messages. (Previously, incomplete escape sequences could have caused interference or the text might have been unreadable.)
If checking the session window height, don't apply the height limit if the
window size wouldn't otherwise be changed as a result of Wimp_OpenWindow.
If the window size is changed and, unless Ctrl pressed, > max height, adjust
the vertical scroll bar such that if it was already at the bottom it remains
there.
Perhaps make this scrollbar handling configurable?