Fix for the validation string parser failing to cope with spaces, and to support escaping within the name.
When parsing the validation string, the search (and copy) would terminate
on when anything greater than a space was seen. This meant that any time
that there was a space in a validation string (for example
'A~ ;Nsomething' to disallow spaces in an input field) the search would
fail before it found the subsequent name. I've merely changed the check
to allow spaces as well.

Escape characters can now be included within the name to allow a ; to be
used in the name. Obviously this won't matter for this parser as the name
should never contain a ; - it would fail if so because that's not a valid
identifier character for C.  But purely to be correct I've added.

Changes are feedback from template checking tool which is very loosely
based on this source.