Update release scripts for beta release.
Update various dates from 2003/2004 to 2010.

Update ZapRedraw to 0.48 for BeagleBoard compatibility - this also
requires a 'fake' ZapFontsMessages path to be configured in the event
that the user doesn't have a ZapFonts installation, so have added these
messages too, and configuration of them into the !Run file.

Change the build options to add -c90 (this seems to fix the __assert2
issue that Castle's C compiler causes with stubsg), -Wp to suppress the
non-ANSI include warnings that TCPIPLibs generates, and -memaccess
-L22-S22-L41 to build the binary in a way that the BeagleBoard should be
happy with. (These changes are partly based on the ones that James
Peacock sent me).

And do a nasty change to c/choices to allow the errno.h that I have to
build with -c90 - it appears to fail to build for me, so temporarily
disable the #include and locally declare errno.

Tested/built on RPCEmu and it seems to run/display OK.